Split large html/ files into focused modules
Pure file moves — no logic changes.

css.go (1104 → 730 lines):
  Extracted selector matching, pseudo-class evaluation, and DOM
  traversal helpers into css_selectors.go (387 lines). Includes
  selectorMatches, partMatches, pseudoMatches, all :nth-*()
  functions, An+B parsing, and node attribute helpers.

converter_style.go (1231 → 948 lines):
  Extracted standalone CSS value parsers into
  converter_style_parsers.go (299 lines). Includes parseTransform,
  parseAngle, parseBoxShadow, parseBoxShadows, splitTopLevelCommas,
  parseTransformOrigin, and numeric/length helpers.
